data validation

+++ hash_all.py
@@ -1,6 +1,7 @@
 import hashlib
 import os
 import sys
+import time
 from pathlib import Path
 
 def md5(fname):
@@ -10,44 +11,59 @@ def md5(fname):
             hash_md5.update(chunk)
     return hash_md5.hexdigest()
 
-def walk (dirr, hash_loc):
+def printt(message):
+    terminal.write(message+"\n")
+    log.write(message+"\n")
 
-    os.makedirs(hash_loc, exist_ok=True)
+def walk (dirr, hash_loc, read_only = False):
 
-    print (dirr)
+    if not read_only:
+        os.makedirs(hash_loc, exist_ok=True)
 
-    for filename in os.listdir(dirr):
-        full_path = os.path.join(dirr,filename)
+    printt (dirr)
+
+    for f1 in os.listdir(dirr):
+        full_path = os.path.join(dirr,f1)
         if os.path.isdir(full_path):
-            if "metadata_hash" not in filename:
-                walk(full_path, os.path.join (hash_loc, filename))
+            if "metadata_hash" not in f1:
+                walk(full_path, os.path.join (hash_loc, f1))
         else:
-            hash_file = os.path.join(hash_loc,filename+".json")
+            hash_file = os.path.join(hash_loc,f1+".json")
             if os.path.exists(hash_file):
                 try:
                     with open(hash_file, "r") as hash_json:
                         last_hash = hash_json.readline()
                         md5_string = md5(full_path)
                         if last_hash != md5_string:
-                            print(f"hash error {hash_file}", file=sys.stderr)
-                except _:
-                    print(f"error reading hash {hash_file}", file=sys.stderr)
+                            printt(f"hash error {hash_file}")
+                except:
+                    printt(f"error reading hash {hash_file}")
             else:
-                with open(hash_file, "w") as hash_json:
-                    md5_string = md5(full_path)
-                    print (f"new file. adding hash. {filename} : {md5_string}" ,file=sys.stderr)
-                    hash_json.write(md5_string)
+                if not read_only:
+                    with open(hash_file, "w") as hash_json:
+                        md5_string = md5(full_path)
+                        printt (f"new file. adding hash. {f1} : {md5_string}")
+                        hash_json.write(md5_string)
 
     # for every hash file there should
-    for filename in os.listdir(hash_loc):
-        if not os.path.isdir(os.path.join (hash_loc, filename)) and ".json" in filename:
-            original_filename = os.path.join(dirr, filename[:-5])
+    for f2 in os.listdir(hash_loc):
+        if not os.path.isdir(os.path.join (hash_loc, f2)) and ".json" in f2:
+            original_filename = os.path.join(dirr, f2[:-5])
             if not os.path.exists(original_filename):
-                print(f"hash exists for missing file {filename}", file=sys.stderr)
+                printt(f"hash exists for missing file {f2}")
 
+
+print("starting hash run.")
 
 hash_root = "./metadata_hashes"
 
+read_only = len ( sys.argv ) == 1
+
+terminal = sys.stdout
+log = open(f"hash_result{int(time.time())}.log", "a")
+
 for filename in os.listdir("."):
     if os.path.isdir(filename) and "metadata_hashes" not in filename and "metadata_website" not in filename:
-        walk(os.path.join (".", filename), os.path.join (hash_root, filename))
+        walk(os.path.join (".", filename), os.path.join (hash_root, filename), read_only)
+
+print("hash run complete.")
